requirements.txt
beautifulsoup4==4.10.0
boto3==1.20.49
botocore==1.23.49
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
dnspython==2.2.0
idna==3.3
jmespath==0.10.0
numpy==1.22.2
pymongo==4.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.8
